Refactored mailer functionality into a module.
Enables easily extending third party mailers as devise mailers.
